Various changes
 - Purge puppet on base and nocm boxes more effectively.
 - Handle unavailable tools better.
 - Get hostname detection working.
 - Provision less memory, 1 CPU, and thin disks.
 - Change guest_os_type to one vSphere actually understands.

+++ scripts/cleanup-puppet.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+# Run a command only if it exists
+if_exists () {
+  command -v $1 >/dev/null && "$@"
+}
+
 # PE is currently used to provision nocm and puppet boxes.
 # This script cleans up directories that may be left around
 # as part of that process.
@@ -9,10 +14,14 @@ if [ -n "${PUPPET_NFS}" ]; then
   umount -l /opt/puppet
 fi
 
-# Only remove /etc/puppetlabs on -nocm boxes
+# Remove /etc/puppetlabs on -nocm boxes
 case "${PACKER_BUILD_NAME}" in *-nocm)
-  rm -rf /etc/puppetlabs
+  NOCM=1
 esac
+
+if [ -n "$NOCM" ] ; then
+  rm -rf /etc/puppetlabs
+fi
 
 # Run the PE uninstaller on Amazon builders
 case "${PACKER_BUILDER_TYPE}" in amazon-*)
@@ -20,8 +29,12 @@ case "${PACKER_BUILDER_TYPE}" in amazon-*)
   ./puppet-enterprise-uninstaller -d -p -y
 esac
 
+echo "Making sure there aren't any Puppet packages installed"
+if_exists apt-get -y --purge remove puppet-agent || true
+if_exists yum remove package puppet-agent || true
+
 # Remove other Puppet-related files and directories
-rm -rf /opt/puppet /opt/puppetlabs /opt/puppet-agent
+rm -rf /opt/puppet*
 rm -rf /var/cache/yum/puppetdeps
 rm -rf /var/opt/lib/pe-puppet
 rm -rf /var/opt/puppetlabs
